refactor: ImageSlider IndexIndicator 컴포넌트 분리
