+++ src/components/atoms/ImageSlider.tsx
@@ -2,12 +2,9 @@ import { ReactNode, useState } from 'react';
 
 import Slider from 'react-slick';
 
-import '@/slick.css';
+import IndexIndicator from './IndexIndicator';
 
-type IndexIndicatorProps = {
-	currentIndex: number;
-	totalImageNumber: number;
-};
+import '@/slick.css';
 
 type ImageSliderProps = {
 	children: ReactNode;
@@ -24,23 +21,6 @@ const initOptions = {
 	dots: false,
 	autoplay: false,
 	infinite: false,
-};
-
-const IndexIndicator = ({
-	currentIndex,
-	totalImageNumber,
-}: IndexIndicatorProps) => {
-	return (
-		<div className="flex items-center gap-1 px-3 py-1 rounded-[65px] bg-black opacity-60">
-			<h6 className="text-White text-center text-xs font-semibold ">
-				{currentIndex}
-			</h6>
-			<div className="w-0.5 h-2 bg-Gray20" />
-			<h6 className="text-Gray20 text-center typography-Caption1">
-				{totalImageNumber}
-			</h6>
-		</div>
-	);
 };
 
 const ImageSlider = ({
